+++ str_len_func.py
